refactor: update allowlist util
