+++ app/components/ClaimShareDrawer/ClaimShareDrawer.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { isUserInAllowlist } from "@/app/actions/claim";
 import { TRACKING_EVENTS } from "@/app/constants/tracking";
 import { useToast } from "@/app/providers/ToastProvider";
 import { copyTextToClipboard } from "@/app/utils/clipboard";
@@ -37,7 +36,6 @@ const ClaimShareDrawer = ({
   const { infoToast } = useToast();
   const [ogImageUrl, setOgImageUrl] = useState("");
   const [showMysteryBox, setShowMysteryBox] = useState(false);
-  const [isUserWhitelisted, setIsUserWhitelisted] = useState(false);
 
   const handleCopy = async () => {
     await copyTextToClipboard(copyUrl);
@@ -69,18 +67,9 @@ const ClaimShareDrawer = ({
     if (!!copyUrl) fetchLinkPreview();
   }, [isOpen, copyUrl]);
 
-  const checkAllowlist = async () => {
-    const isAllowlisted = await isUserInAllowlist();
-    setIsUserWhitelisted(isAllowlisted);
-  };
-
-  useEffect(() => {
-    checkAllowlist();
-  }, []);
-
   return (
     <>
-      {isUserWhitelisted && showMysteryBox && mysteryBoxId ? (
+      {showMysteryBox && mysteryBoxId ? (
         <MysteryBox
           isOpen={showMysteryBox}
           closeBoxDialog={() => {
